feat:TOP-71 Hold를 할 때 시간초가 정수단위로 롤백되는 현상 수정
 - Timer를 정수단위에서 0.5단위로 동작하도록 해 Hold on/off시에 발생되는 간격 차를 최소화

+++ feature/tohot/src/main/java/tht/feature/tohot/component/card/ToHotCard.kt
@@ -43,13 +43,13 @@ fun ToHotCard(
     introduce: String,
     timer: CardTimerUiModel.ToHotTimer,
     maxTimeSec: Int,
-    currentSec: Int,
-    destinationSec: Int,
+    currentSec: Float,
+    destinationSec: Float,
     enable: Boolean,
     fallingAnimationEnable: Boolean = false,
     isHoldCard: Boolean,
     onFallingAnimationFinish: () -> Unit = { },
-    ticChanged: (Int) -> Unit = { },
+    ticChanged: (Float) -> Unit = { },
     userCardClick: () -> Unit = { },
     onLikeClick: () -> Unit = { },
     onUnLikeClick: () -> Unit = { },
@@ -168,8 +168,8 @@ private fun ToHotCardPreview() {
         introduce = "introduce",
         timer = CardTimerUiModel.ToHotTimer.Timer,
         maxTimeSec = 5,
-        currentSec = 5,
-        destinationSec = 4,
+        currentSec = 5f,
+        destinationSec = 4f,
         enable = true,
         isHoldCard = false
     )
@@ -195,8 +195,8 @@ private fun ToHotCardHoldCardPreview() {
         introduce = "introduce",
         timer = CardTimerUiModel.ToHotTimer.Timer,
         maxTimeSec = 5,
-        currentSec = 5,
-        destinationSec = 4,
+        currentSec = 5f,
+        destinationSec = 4f,
         enable = true,
         isHoldCard = true
     )
@@ -222,8 +222,8 @@ private fun ToHotHeartCardPreview() {
         introduce = "introduce",
         timer = CardTimerUiModel.ToHotTimer.Heart,
         maxTimeSec = 5,
-        currentSec = 5,
-        destinationSec = 4,
+        currentSec = 5f,
+        destinationSec = 4f,
         enable = true,
         isHoldCard = false
     )
@@ -249,8 +249,8 @@ private fun ToHotDislikeCardPreview() {
         introduce = "introduce",
         timer = CardTimerUiModel.ToHotTimer.Dislike,
         maxTimeSec = 5,
-        currentSec = 5,
-        destinationSec = 4,
+        currentSec = 5f,
+        destinationSec = 4f,
         enable = true,
         isHoldCard = false
     )
